transfer Bot class to bot.py file
